Extracted superclasses with common elements from XTA propagation analysis.
